Optimization with Quassi Newton
This Python script implements the Quassi Newton algorithm to find the minimum of a two-variable mathematical function. The starting point, function and other parameters can be changed optionally. The code is an instructive example for those who want to solve numerical optimization problems.

Note: This script was written for numerical optimization and learning on the Quassi Newton algorithm.